Got 1st part of the program to work
In Executive.cpp I manually inputted the file needed. That can be seen as input.txt. Will write to the intermediate.txt
FileManagement.h and .cpp have been updated to be able to work to write and read from files. Thinking we can just call those and not have to worry about accessing the files in the other classes. 
Workflow.h and .cpp has been updated to go through the first part of the program. 
Map.h and .cpp has been set to be given the key(file to write too) and the line of raw data to convert. Still not sure if this is what the prof. intended, very hard to understand the instructions, still reaching out. Let me know if you think its something else. 
Still have to clean everything up, but thought this was a good stopping point for now.