Merged both programs
Got both programs to work. I am going to try and move some of the commands in the while loop you made to the FileManagement. Just so it looks cleaner on the workflow. 
I did a test with 20 paragraphs and results look promising. (Those can be seen in the test2.txt files). I did notice one bug that the tokenizer will not pick up any words that have accent over the letter. It also will take words like it's and make it ("it") 1 and ("s") 1.