Code cleanup and error scenarios
I adjusted one part of the Reduce.cpp class because in the instructions he stated "The reduce method itself should not directly deal with any File IO", so I just made it so it calls the FileManagement for the inserting of the output. 
I also started working on some error scenarios and using try and catch to correct them. So far it works to catch if a bad input file is given. 
It keeps crashing when trying to check the directories of the temp and output files. 
Updated so a 'SUCCESS' file gets written to the output directory when program is finished.
Also made it possible for the program to be able to get file directories even if they are not local to the program.