Copy over everything-Added functionality
File dump. Feel like we are in a solid place. 
I was able to copy all your notes over I believe, apologize if something got left out. 
I was able to make it so that if just a directory is given as the input file, then it will search for all files in the directory and put it in one output file. 
If just a file is given, then it will just do the one file. 
Also added a stop watch to the main() to time how long the program takes. Thought that might be good to have and maybe expanded to see how long each part of the program is taking. 
I do still want to try a function that will prompt for a different file if the name is not correct or file not found. 
After testing the shakespeare directory, I do see that the bug is present when a accent char is given. Got to find a way around that.