Add check for file name, Add Boost logging
I put in a check to make sure user is giving valid log names. I am not 100% if it will write directories if you don't input files that do not exist. I was also able to get the boost logging going to where it writes the error codes to a file sample.log(Note:this file only exist once the program is complete). I worked on replacing some of the catach parts of the try-catch blocks with this logging. So now it will print to the log instead of showing on screen. Let me know what you think on that, I just tried the workflow.cpp file in case you had another idea.