Reupload program
Reuploaded the program so we both are on the same path. Everything should work. Test.cpp file should be added-BOOST unit test. copied and pasted your sorting program to the inputIsFile function. You had only put it in the directory function. There is a function depending on if the input is just a file or a directory. Also added some more get functions so that they can be tested in workflow.