Initial update
So this was my main idea, so can use as big of an R as you like, just have to put it into const RMAX. Believe it is at 100 right now, so just watch that if your going to test because it will create 100 temp files. Just a start, still have a few things want to work on, like to clear files on the first pass. But hoprfully can get you started on something as well.