Put Reduce Process in one file
Put these in one file. I did make one change to your main.cpp. I noticed it was breaking when user input a directory tree that had a space in it. So the update basically converts any spaces to '\n' in the MapReduce workflow and the main in the processes converts it back to a space. 
Just any easy fix, let me know if you have any more efficient idea.