Updated Workflow
Added a partition function that is called to do the processes. I did add one little thing to the part 2 section, because I noticed it was having a issue if you had any space in directory tree that is given. So it simply just replaces the spaces with '\n' and converts it back when it gets to the process. 
So far this works for same directory and small tests. 
However, It sort of breaks down when dealing with a large directory like shakespeare. I noticed that it is trying to do both the map and the reduce at the same time almost, because workflow is not waiting for the processes to finish. I put a sleep to try and slow it down, but still had an issue. I think if another sleep is added to the end, it might give it time to finish. Just wanted to get you an update so we can try to brainstorm ideas. 
Let me know.