Add files via upload
Found a bug that it would delete all the words that had previously been completed and then move on to the next thread. This caused the output to only have the counts for the last file it read in. Deleted the part of Map Process that was clearing out the threads. In the new workflow.cpp I added a for loop to clear out the temp files that get created so it still gets cleaned up.